+++ kohese-portal/server/kdb.js
@@ -190,17 +190,14 @@ module.exports.retrieveModelInstance = retrieveModelInstance;
 //////////////////////////////////////////////////////////////////////////
 //
 //////////////////////////////////////////////////////////////////////////
-function storeModelInstance(modelName, modelInstance, isNewItem){
+function storeModelInstance(proxy, isNewItem){
 
-
-  var proxy = ItemProxy.getProxyFor(modelInstance.id);
+  var modelName = proxy.kind;
+  var modelInstance = proxy.item;
   
   if(modelName !== "Analysis"){
     // Delete any associated analysis
-    var analysisStore = kdbStore.models["Analysis"];
-    if(analysisStore[modelInstance.id]){
-      removeModelInstance("Analysis", modelInstance.id);
-    }
+    removeModelAnalysis(proxy);
   }
 
   var repo = proxy.getRepositoryProxy();    
@@ -272,8 +269,35 @@ module.exports.storeModelInstance = storeModelInstance;
 //////////////////////////////////////////////////////////////////////////
 //
 //////////////////////////////////////////////////////////////////////////
-function removeModelInstance(modelName, instanceId){
-  var proxy = ItemProxy.getProxyFor(instanceId);
+function storeModelAnalysis(analysisInstance){
+  var modelName = 'Analysis';
+  var modelInstanceId = analysisInstance.id;
+  console.log(analysisInstance);
+  
+  var proxy = ItemProxy.getProxyFor(modelInstanceId);
+  
+  if (proxy) {
+    var repo = proxy.getRepositoryProxy();
+    var analysisPath = determineRepoStoragePath(repo) + '/Analysis/' + modelInstanceId + '.json';
+    
+    var modelStore = kdbStore.models[modelName];
+    modelStore[modelInstanceId] = JSON.stringify(analysisInstance);
+    
+    kdbFS.storeJSONDoc(analysisPath, analysisInstance);
+
+    
+  } else {
+    console.log('*** No item found for analysis instance: ' + modelInstanceId);
+  }
+}
+module.exports.storeModelAnalysis = storeModelAnalysis;
+
+//////////////////////////////////////////////////////////////////////////
+//
+//////////////////////////////////////////////////////////////////////////
+function removeModelInstance(proxy){
+  var modelName = proxy.kind;
+  var instanceId = proxy.item.id;
 
   if(modelName === "Analysis"){
     var repo = proxy.getRepositoryProxy();    
@@ -283,18 +307,14 @@ function removeModelInstance(modelName, instanceId){
     kdbFS.removeFile(filePath); 
   } else if (modelName === 'Repository') {
       if(mountList[proxy.item.id].mounted) {
-       // Delete any associated analysis
-          var analysisStore = kdbStore.models["Analysis"];
-          if(analysisStore[instanceId]){
-            removeModelInstance("Analysis", instanceId);
-          }
+          // Delete any associated analysis
+          removeModelAnalysis(proxy);
           
           var parentRepo = proxy.parentProxy.getRepositoryProxy();
           var parentRepoPath = mountList[parentRepo.item.id].repoStoragePath;
           
           kdbFS.removeFile(path.join(parentRepoPath, 'Repository', proxy.item.id + '.json.mount'));
           delete mountList[proxy.item.id];
-          proxy.deleteItem();
           updateMountFile();
           
           // Should remove the proxies and modelStore instances of all the children, but not delete from disk.
@@ -304,25 +324,46 @@ function removeModelInstance(modelName, instanceId){
           // In that case, remove it from the mount and remove the proxy.
           
           delete mountList[proxy.item.id];
-          proxy.deleteItem();
           updateMountFile();
       }
   } else {
     // Delete any associated analysis
-    var analysisStore = kdbStore.models["Analysis"];
-    if(analysisStore[instanceId]){
-      removeModelInstance("Analysis", instanceId);
-    }
+    removeModelAnalysis(proxy);
 
-    kdbFS.removeFile(proxy.repoPath);
-    proxy.deleteItem();
+    if (proxy.repoPath){
+      kdbFS.removeFile(proxy.repoPath);
+      delete proxy.repoPath;
+    }
   }
   
   var modelStore = kdbStore.models[modelName];
-  delete modelStore[instanceId];
+  if(modelStore && modelStore[instanceId]){
+    delete modelStore[instanceId];    
+  }
 }
 
 module.exports.removeModelInstance = removeModelInstance;
+
+
+//////////////////////////////////////////////////////////////////////////
+//
+//////////////////////////////////////////////////////////////////////////
+function removeModelAnalysis(proxy) {
+  var modelInstanceId = proxy.item.id;
+  
+  var analysisStore = kdbStore.models["Analysis"];
+  if(analysisStore[modelInstanceId]){
+    delete analysisStore[modelInstanceId];
+  }
+
+  var repo = proxy.getRepositoryProxy();
+  var analysisPath = determineRepoStoragePath(repo) + '/Analysis/' + modelInstanceId + '.json';
+  
+  const ignoreNonExistent = true;
+  kdbFS.removeFile(analysisPath, ignoreNonExistent);
+  
+}
+module.exports.removeModelAnalysis = removeModelAnalysis;
 
 //////////////////////////////////////////////////////////////////////////
 //
